RDCC-958: moving Enum classes to Constants package, adding test for ResourceNotFoundException
