+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/constants/IdamStatus.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.professionalapi.controller.constants;
+
+public enum IdamStatus {
+    ACTIVE,PENDING,SUSPENDED
+}
