+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/constants/PrdEnumType.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.professionalapi.service;
+package uk.gov.hmcts.reform.professionalapi.controller.constants;
 
 public enum PrdEnumType {
     SIDAM_ROLE, ADMIN_ROLE, JURISD_ID
