--- src/main/java/uk/gov/hmcts/reform/professionalapi/controller/response/IdamStatus.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.professionalapi.controller.response;
-
-public enum IdamStatus {
-    ACTIVE,PENDING,SUSPENDED
-}
