+++ src/main/java/uk/gov/hmcts/reform/professionalapi/domain/ProfessionalUser.java
@@ -26,7 +26,7 @@ import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import uk.gov.hmcts.reform.professionalapi.controller.response.IdamStatus;
+import uk.gov.hmcts.reform.professionalapi.controller.constants.IdamStatus;
 
 @Entity(name = "professional_user")
 @NoArgsConstructor
