+++ src/main/java/uk/gov/hmcts/reform/professionalapi/domain/UserProfile.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.professionalapi.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import uk.gov.hmcts.reform.professionalapi.controller.response.IdamStatus;
+import uk.gov.hmcts.reform.professionalapi.controller.constants.IdamStatus;
 
 @Getter
 @AllArgsConstructor
