+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/request/validator/ProfessionalUserReqValidatorTest.java
@@ -8,9 +8,8 @@ import java.util.UUID;
 
 import org.junit.Before;
 import org.junit.Test;
+import uk.gov.hmcts.reform.professionalapi.controller.constants.IdamStatus;
 import uk.gov.hmcts.reform.professionalapi.controller.request.InvalidRequest;
-import uk.gov.hmcts.reform.professionalapi.controller.request.validator.ProfessionalUserReqValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.response.IdamStatus;
 import uk.gov.hmcts.reform.professionalapi.domain.RoleName;
 import uk.gov.hmcts.reform.professionalapi.domain.UserProfileUpdatedData;
