+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/response/GetUserProfileResponseTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.UUID;
 import org.junit.Test;
 
+import uk.gov.hmcts.reform.professionalapi.controller.constants.IdamStatus;
 import uk.gov.hmcts.reform.professionalapi.domain.UserProfile;
 
 public class GetUserProfileResponseTest {
