+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/response/MultipleUserProfilesResponseTest.java
@@ -8,6 +8,7 @@ import java.util.UUID;
 
 import org.junit.Before;
 import org.junit.Test;
+import uk.gov.hmcts.reform.professionalapi.controller.constants.IdamStatus;
 import uk.gov.hmcts.reform.professionalapi.domain.UserProfile;
 
 public class MultipleUserProfilesResponseTest {
